requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.7
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pylibdmtx==0.1.10
pyzbar==0.1.9
requests==2.32.3
urllib3==2.2.2
